refactor: Update date formatting in components
